zlib: fix conversion warnings

+++ libs/zlib/src/crc32.c
@@ -470,7 +470,7 @@ ZLIB_INTERNAL void crc_reset(deflate_state *const s) {
     s->strm->adler = crc32(0L, NULL, 0);
 }
 
-ZLIB_INTERNAL void copy_with_crc(z_stream *strm, uint8_t *dst, uint64_t size) {
+ZLIB_INTERNAL void copy_with_crc(z_stream *strm, uint8_t *dst, size_t size) {
     zmemcpy(dst, strm->next_in, size);
     strm->adler = crc32(strm->adler, dst, size);
 }
